requirements.txt
numpy~=1.26.1
Pillow~=10.0.1
Cython~=3.0.4
numba~=0.58.1